Optional platform headers moved where they ought to be
Peripheral headers that are used indirectly when configuring project
shouldn't be part of the platform mandatory interface.
