+++ platform/stm32f4xx/console.cpp
