+++ platform/stm32f4xx/export/aux/i2c_bus.hpp
