+++ platform/stm32f4xx/export/aux/usart_bus.hpp
