+++ platform/stm32l1xx/templates/console.hpp.in
@@ -5,7 +5,7 @@
 #ifndef STM32L1XX_CONSOLE_DRIVER_HPP_
 #define STM32L1XX_CONSOLE_DRIVER_HPP_
 
-#include <platform/usart_bus.hpp>
+#include <aux/usart_bus.hpp>
 
 namespace ecl
 {
